Added level chart
It fills based on your progress to the next level, and is on the same level as cool as the Steam DB site. I have the infrastructure for badges now as well, I just need to implement it.

I also switched the PHP file over to a switch case, because it's cleaner than the same `==` statements repeated over and over to check the value of one condition.